Add files via upload
This project focuses on conducting exploratory data analysis (EDA) of global layoffs using MySQL. The dataset used in this analysis contains information about layoffs from various companies across different industries and regions. The goal of the analysis is to identify trends, patterns, and outliers in the data, providing insights into the impact of layoffs on businesses worldwide.

The analysis includes several steps, such as:

Cleaning and preprocessing the data to ensure consistency and accuracy.
Identifying companies with the largest single-day layoffs and those with the most cumulative layoffs over time.
Analyzing layoffs by location, industry, and company stage to understand geographical and sector-specific trends.
Examining the severity of layoffs as a percentage of total employees.
Identifying top companies with the most layoffs each year to understand recurring patterns.
The project utilizes SQL queries to extract, transform, and analyze the data stored in a MySQL database. It also includes comments within the code to explain each step of the analysis and provide context for the queries used.

This project serves as a demonstration of EDA techniques using SQL, providing valuable insights for businesses, researchers, and policymakers interested in understanding the dynamics of layoffs on a global scale.